requirements/base.txt
Django==3.1.3
djangorestframework==3.12.2
psycopg2==2.8.6
Authlib==0.15.2
requests==2.24.0